requirements.txt
requests
beautifulsoup4
pytesseract
pillow
